Renamed FileReader to DataLoader.
This better clarifies the distinction between this class, which reads data from
a file and converts it to numpy representation for analysis, and MixedLogReader
(which DataLoader uses under the hood), which is a lower-level class for loading
FusionEngine messages from a file.
